+++ src/r0/send.rs
@@ -1,5 +1,0 @@
-//! Endpoints for sending events.
-
-pub mod send_message_event;
-pub mod send_state_event_for_empty_key;
-pub mod send_state_event_for_key;