+++ src/r0/sync.rs
@@ -1,8 +1,3 @@
 //! Endpoints for getting and synchronizing events.
 
-pub mod get_member_events;
-pub mod get_message_events;
-pub mod get_state_events;
-pub mod get_state_events_for_empty_key;
-pub mod get_state_events_for_key;
 pub mod sync_events;